Add files via upload
About
I have taken the #100DaysOfCodeChallenge, committing to coding for at least an hour every day for 100 days. I have chosen this challenge to improve my coding skills and build a habit of consistent coding. I will be documenting my progress on this GitHub repository and sharing it on Linkedln using the hashtag #100daysofcodechallenge.