Create codechef4
Team RCB has earned 
�
X points in the games it has played so far in this year's IPL. To qualify for the playoffs they must earn at least a total of 
�
Y points. They currently have 
�
Z games left, in each game they earn 
2
2 points for a win, 
1
1 point for a draw, and no points for a loss.

Is it possible for RCB to qualify for the playoffs this year?