Create code_chef _11
Chef wants to appear in a competitive exam. To take the exam, there are following requirements:

Minimum age limit is 
�
X (i.e. Age should be greater than or equal to 
�
X).
Age should be strictly less than 
�
Y.
Chef's current Age is 
�
A. Find whether he is currently eligible to take the exam or not.